pyproject.toml
[tool.poetry]
name = "powerline-podman"
version = "1.0.3"
description = "A custom Powerline segment for displaying the current state of Podman containers"
authors = ["Jordan Duabe <me@jduabe.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/j4ckofalltrades/powerline-podman"
repository = "https://github.com/j4ckofalltrades/powerline-podman"
keywords = ["powerline", "podman"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/j4ckofalltrades/powerline-podman/issues"

[tool.poetry.dependencies]
python = "^3.9"
podman = "^4.6.0"
powerline-status = "^2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
